+++ src/forms/HotkeySettings.ts
@@ -1,10 +1,13 @@
 import ARCHITECT from "../core/architect.js";
 import KEYMAP from "../_data/keymap.js";
-import { HotSwap, KeyMap, LayerShortcuts } from "../general/LayerShortcuts.js";
+import { HotSwap, LayerShortcuts } from "../general/LayerShortcuts.js";
+import { WallCtrlInvert } from "../walls/WallCtrlInvert.js";
+import { KeyMap } from '../core/hotkeys.js';
 
 
 interface RenderOptions {
 	hotswap: HotSwap,
+	wallCtrlInvert: KeyMap,
 	layers: {
 		name: string,
 		label: string,
@@ -16,7 +19,7 @@ interface RenderOptions {
 	}[]
 }
 export default class HotkeyConfig extends FormApplication<RenderOptions> {
-	private static readonly PREF_MENU = "LayerShortcutsSettingsMenu";
+	private static readonly PREF_MENU = "HotkeySettingsMenu";
 	static get defaultOptions(): FormApplication.Options {
 		return mergeObject(super.defaultOptions, {
 			title: 'DF_ARCHITECT.ModName',
@@ -26,7 +29,7 @@ export default class HotkeyConfig extends FormApplication<RenderOptions> {
 			submitOnClose: false,
 			closeOnSubmit: true,
 			width: 525,
-			id: 'DFArchLSMenu',
+			id: 'DFArchHKMenu',
 			template: 'modules/df-architect/templates/LayerShortcutsSettings.hbs'
 		});
 	}
@@ -43,6 +46,7 @@ export default class HotkeyConfig extends FormApplication<RenderOptions> {
 	getData(options?: Application.RenderOptions): RenderOptions {
 		const result: RenderOptions = {
 			hotswap: LayerShortcuts.hotSwap,
+			wallCtrlInvert: WallCtrlInvert.hotkey,
 			layers: ui.controls.controls.map(x => {
 				const result = {
 					name: x.name,
@@ -60,12 +64,14 @@ export default class HotkeyConfig extends FormApplication<RenderOptions> {
 		if (!formData) return;
 		const data = expandObject(formData) as {
 			hotswap: HotSwap,
+			wallCtrlInvert: KeyMap,
 			layers: Indexable<KeyMap>
 		};
 		ui.controls.controls.forEach(layer => {
 			LayerShortcuts.setLayerSetting(layer.name, data.layers[layer.name]);
 		});
 		LayerShortcuts.hotSwap = data.hotswap;
+		WallCtrlInvert.hotkey = data.wallCtrlInvert;
 		ARCHITECT.requestReload();
 	}
 
@@ -75,18 +81,27 @@ export default class HotkeyConfig extends FormApplication<RenderOptions> {
 			e.preventDefault();
 			for (let layer of ui.controls.controls) {
 				const defValue: KeyMap = LayerShortcuts.getLayerDefault(layer.name);
-				$(`#DFArchLSMenu select[name="layers.${layer.name}.key"]`).val(defValue.key);
-				($(`#DFArchLSMenu input[name="layers.${layer.name}.alt"]`)[0] as HTMLInputElement).checked = defValue.alt;
-				($(`#DFArchLSMenu input[name="layers.${layer.name}.ctrl"]`)[0] as HTMLInputElement).checked = defValue.ctrl;
-				($(`#DFArchLSMenu input[name="layers.${layer.name}.shift"]`)[0] as HTMLInputElement).checked = defValue.shift;
+				$(`#DFArchHKMenu select[name="layers.${layer.name}.key"]`).val(defValue.key);
+				($(`#DFArchHKMenu input[name="layers.${layer.name}.alt"]`)[0] as HTMLInputElement).checked = defValue.alt;
+				($(`#DFArchHKMenu input[name="layers.${layer.name}.ctrl"]`)[0] as HTMLInputElement).checked = defValue.ctrl;
+				($(`#DFArchHKMenu input[name="layers.${layer.name}.shift"]`)[0] as HTMLInputElement).checked = defValue.shift;
 			}
-			const defValue: HotSwap = LayerShortcuts.hotSwapDefault;
-			$(`#DFArchLSMenu select[name="hotswap.layer1"]`).val(defValue.layer1);
-			$(`#DFArchLSMenu select[name="hotswap.layer2"]`).val(defValue.layer2);
-			$(`#DFArchLSMenu select[name="hotswap.key"]`).val(defValue.map.key);
-			($(`#DFArchLSMenu input[name="hotswap.alt"]`)[0] as HTMLInputElement).checked = defValue.map.alt;
-			($(`#DFArchLSMenu input[name="hotswap.ctrl"]`)[0] as HTMLInputElement).checked = defValue.map.ctrl;
-			($(`#DFArchLSMenu input[name="hotswap.shift"]`)[0] as HTMLInputElement).checked = defValue.map.shift;
+			{
+				const defValue: HotSwap = LayerShortcuts.hotSwapDefault;
+				$('#DFArchHKMenu select[name="hotswap.layer1"]').val(defValue.layer1);
+				$('#DFArchHKMenu select[name="hotswap.layer2"]').val(defValue.layer2);
+				$('#DFArchHKMenu select[name="hotswap.map.key"]').val(defValue.map.key);
+				($('#DFArchHKMenu input[name="hotswap.map.alt"]')[0] as HTMLInputElement).checked = defValue.map.alt;
+				($('#DFArchHKMenu input[name="hotswap.map.ctrl"]')[0] as HTMLInputElement).checked = defValue.map.ctrl;
+				($('#DFArchHKMenu input[name="hotswap.map.shift"]')[0] as HTMLInputElement).checked = defValue.map.shift;
+			}
+			{
+				const defValue: KeyMap = WallCtrlInvert.hotkeyDefault;
+				$('#DFArchHKMenu select[name="wallCtrlInvert.key"]').val(defValue.key);
+				($('#DFArchHKMenu input[name="wallCtrlInvert.alt"]')[0] as HTMLInputElement).checked = defValue.alt;
+				($('#DFArchHKMenu input[name="wallCtrlInvert.ctrl"]')[0] as HTMLInputElement).checked = defValue.ctrl;
+				($('#DFArchHKMenu input[name="wallCtrlInvert.shift"]')[0] as HTMLInputElement).checked = defValue.shift;
+			}
 		});
 	}
 }