+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # DF Architect
 
 ## Alpha Release v0.3.0 (2021-03-13)
-- Layer Hot Swap shortcut. You can select 2 layers that a hotkey can quickly swap between.
+- Layer Hot Swap shortcut. You can select 2 layers that a hotkey can quickly swap between. Defaulted to `Alt + Q`.
+- Added Wall Chain Lock. This essentially inverts the Ctrl key behaviour.
+- Added Wall Chain Lock hotkey, defaulted to `Alt + C`.
 
 ## Alpha Release v0.2.0 (2021-03-07)
 - Alternate Grid Snapping now works perfectly.